rename instance_config model -> controller
