+++ src/apps/meta/pages/fix_personalised_instance_id_record.vue
@@ -32,7 +32,6 @@
 </template>
 
 <script>
-  //import { mapState, mapActions, mapMutations } from 'vuex'
   import virtual_list from 'vue-virtual-scroll-list'
   import moment from "moment"
   import {get} from 'lodash'
