+++ src/apps/meta/routes.js
@@ -1,6 +1,5 @@
 import applet from './applet.vue'
 import home from './pages/home.vue'
-import login from './pages/login.vue'
 import logout from './pages/logout.vue'
 import clear_data from './pages/clear_data.vue'
 import geodata from './pages/geodata.vue'
@@ -18,10 +17,6 @@ export default [
         name: 'meta:home',
         component: home,
         meta: {title: 'Home'}
-      },{
-        path: 'login',
-        name: 'meta:login',
-        component: login,
       },{
         path: 'logout',
         name: 'meta:logout',
