+++ src/index.js
@@ -6,10 +6,11 @@ import '@mapbox/mapbox-gl-draw/dist/mapbox-gl-draw.css'
 import 'lib/bootstrap-extract.css' // contains only table-related stuff, but still TODO: remove reliance on bootstrap for monitor-table
 import 'vue-material/dist/vue-material.css'
 import 'vue-multiselect/dist/vue-multiselect.min.css'
+
 // Imports
 import pubsubcache from 'lib/helpers/pubsubcache'
 import 'config/configure_vue'
-import {retrieve_local_config} from '../shell_app/instance_config/model'
+import {retrieve_local_config} from './shell_app/instance_config/controller'
 import {configure_error_tracking} from 'config/error_tracking'
 import {configure_pubsub_converters} from 'config/configure_pubsub_converters'
 import {launch_main_app} from 'config/launch_main_app'
